requirements.txt
Pillow<11,>=10.2.0
install-playwright<2,>=0.0.1
playwright<2,>=1.41.2
